requirements.txt
allure-pytest == 2.15.0
Appium-Python-Client == 5.1.2
pytest == 8.4.1
pytest-xdist == 3.8.0
python-dotenv == 1.1.1
selenium == 4.34.2